Unify test class. Remove unnecesary line @ExtendBy(MockitoExtension.class)
