+++ src/test/java/de/mirkosertic/flightrecorderstarter/core/FlightRecorderTest.java
@@ -2,24 +2,24 @@ package de.mirkosertic.flightrecorderstarter.core;
 
 
 import de.mirkosertic.flightrecorderstarter.FlightRecorderDynamicConfiguration;
+import jdk.jfr.Configuration;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
-import org.junit.jupiter.api.extension.ExtendWith;
-import org.mockito.junit.jupiter.MockitoExtension;
 import org.springframework.boot.system.SystemProperties;
 
 import java.io.IOException;
 import java.nio.file.Files;
 import java.nio.file.Path;
 import java.time.Duration;
+import java.util.ArrayList;
 import java.util.HashMap;
+import java.util.List;
 import java.util.Map;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.mockito.BDDMockito.given;
 import static org.mockito.Mockito.*;
 
-@ExtendWith(MockitoExtension.class)
 class FlightRecorderTest {
 
     private final Map<Long, RecordingSession> spyRecordings;
@@ -71,5 +71,69 @@ class FlightRecorderTest {
         final RecordingSession recordingSession = this.spyRecordings.get(recordingId);
         assertThat(recordingSession).isNotNull();
         assertThat(recordingSession.getRecording().getDestination().getParent()).isEqualTo(defaultTemporalPath);
+    }
+
+
+    @Test
+    void givenConfiguredCustomConfigurationProfile_whenSettingsAreChosen_thenCustomSettingsAreUed() {
+        //Given
+        final List<Configuration> mockConfigurationList = generateMockConfigurationList();
+
+        given(this.mockConfiguration.getJfrCustomConfig()).willReturn("customjfc");
+
+        final FlightRecorder flightRecorder = new FlightRecorder(this.mockConfiguration);
+
+        //When
+        final Map<String, String> mapSettings = flightRecorder.getConfigurationSettings(mockConfigurationList);
+
+        //Then
+        assertThat(mapSettings.get("settingsKey")).isEqualTo("customValue");
+
+
+    }
+
+    @Test
+    void givenNonConfiguredCustomConfigurationProfile_whenSettingsAreChosen_thenProfileSettingsAreUed() {
+        //Given
+        final List<Configuration> mockConfigurationList = generateMockConfigurationList();
+
+        given(this.mockConfiguration.getJfrCustomConfig()).willReturn(null);
+
+        final FlightRecorder flightRecorder = new FlightRecorder(this.mockConfiguration);
+
+        //When
+        final Map<String, String> mapSettings = flightRecorder.getConfigurationSettings(mockConfigurationList);
+
+        //Then
+        assertThat(mapSettings.get("settingsKey")).isEqualTo("profileValue");
+
+
+    }
+
+
+    List<Configuration> generateMockConfigurationList() {
+        final List<Configuration> mockConfigurationList = new ArrayList<>();
+        final Configuration defaultConfiguration = mock(Configuration.class);
+        given(defaultConfiguration.getName()).willReturn("default");
+        final Map<String, String> defaultSettings = new HashMap<>();
+        defaultSettings.put("settingsKey", "defaultValue");
+        given(defaultConfiguration.getSettings()).willReturn(defaultSettings);
+        mockConfigurationList.add(defaultConfiguration);
+
+        final Configuration profileConfiguration = mock(Configuration.class);
+        given(profileConfiguration.getName()).willReturn("profile");
+        final Map<String, String> profileSettings = new HashMap<>();
+        profileSettings.put("settingsKey", "profileValue");
+        given(profileConfiguration.getSettings()).willReturn(profileSettings);
+        mockConfigurationList.add(profileConfiguration);
+
+        final Configuration customConfiguration = mock(Configuration.class);
+        given(customConfiguration.getName()).willReturn("customjfc");
+        final Map<String, String> customSettings = new HashMap<>();
+        customSettings.put("settingsKey", "customValue");
+        given(customConfiguration.getSettings()).willReturn(customSettings);
+        mockConfigurationList.add(customConfiguration);
+
+        return mockConfigurationList;
     }
 }
